requirements.txt
Django
djangorestframework
psycopg2
celery
django-celery-beat
django-channels
channels-redis
drf-yasg